+++ packages/codegen/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.2.3](https://github.com/manifesto-ai/core/compare/codegen-v0.2.2...codegen-v0.2.3) (2026-04-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @manifesto-ai/core bumped to 2.9.0
+  * peerDependencies
+    * @manifesto-ai/core bumped from ~2.0.0 to ~2.9.0
+
+## [0.2.2](https://github.com/manifesto-ai/core/compare/codegen-v0.2.1...codegen-v0.2.2) (2026-04-02)
+
+
+### Features
+
+* **compiler:** emit codegen during dev and rename domain output ([fee61e3](https://github.com/manifesto-ai/core/commit/fee61e35681268e941914f6be2c0374ec98c63c4))
+* **compiler:** emit codegen during dev and rename domain output ([a67b3c0](https://github.com/manifesto-ai/core/commit/a67b3c082923191d3e46c7bbaab7bd117d318c5f))
+
 ## [0.2.1](https://github.com/manifesto-ai/core/compare/codegen-v0.2.0...codegen-v0.2.1) (2026-04-02)